# src/AppForSEII2526.API/Controllers/HerramientasController.cs
using AppForSEII2526.API.DTOs.ComprarDTOs;
using AppForSEII2526.API.DTOs.HerramientaDTO;
using AppForSEII2526.API.DTOs.OfertaDTOs;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using System.Linq;
using static System.Runtime.InteropServices.JavaScript.JSType;

namespace AppForSEII2526.API.Controllers
{
    [Route("api/[controller]")]
    [ApiController]
    public class HerramientasController : ControllerBase
    {
        private readonly ApplicationDbContext _context;
        private readonly ILogger<HerramientasController> _logger;

        public HerramientasController(ApplicationDbContext context, ILogger<HerramientasController> logger)
        {
            _context = context;
            _logger = logger;
        }


        [HttpGet]
        [Route("[action]")]
        [ProducesResponseType(typeof(IList<HerramientaOfertaDTO>), (int)HttpStatusCode.OK)]//Devuelve herramientas filtradas por fabricante y precio maximo

        public async Task<ActionResult> GetSelectFiltradoOferta(string? fabricante, double? precioMaximo = null)
        {

            var query = _context.Herramienta.AsQueryable();

            if (precioMaximo.HasValue)
                query = query.Where(h => h.Precio <= precioMaximo.Value);
            if (!string.IsNullOrEmpty(fabricante))
                query = query.Where(h => h.Fabricante.Nombre.Contains(fabricante));
            var herramientas = await query
                .Select(h => new HerramientaOfertaDTO() {
                    Id = h.Id,
                    Nombre = h.Nombre,
                    Material = h.Material,
                    Precio = (double)h.Precio,
                    Fabricante = h.Fabricante
                })
                .ToListAsync();
            return Ok(herramientas);

        }


        [HttpGet]
        [Route("[action]")]
        [ProducesResponseType(typeof(IList<HerramientaComprarDTO>), (int)HttpStatusCode.OK)]
        [ProducesResponseType((int)HttpStatusCode.NotFound)]
        public async Task<IActionResult> GetHerramientaComprar(string? material, decimal? precio, string? nombreHerramienta)//Devuelve todo lo relativo a Herramienta
        {
            if (_context.Herramienta == null)
            {
                _logger.LogWarning("No se encontraron herramientas");
                return NotFound();
            }
            _logger.LogInformation("Se han encontrado herramientas");

            var herramientas = await _context.Herramienta
                .Include(h => h.Fabricante)
                .Where(h => (material == null || h.Material.ToLower().Contains(material.ToLower())) &&
                       (precio == null || h.Precio <= (double)precio) &&
                       (nombreHerramienta == null || h.Nombre.ToLower().Contains(nombreHerramienta.ToLower())))
                .OrderBy(h => h.Nombre)
                .Select(h => new HerramientaComprarDTO(
                    h.Id,
                    h.Nombre,
                    h.Material,
                    (decimal)h.Precio,
                    h.Fabricante.Nombre
                    ))
                .ToListAsync();
            _logger.LogInformation("Finalizando la muestra de herramientas");
            return Ok(herramientas);
        }


        [HttpGet]
        [Route("[action]")]
        [ProducesResponseType(typeof(IList<HerramientaDTO>), (int)HttpStatusCode.OK)]
        public async Task<IActionResult> GetSelectFiltradoReparacion(string? nombreFiltro, string? tiempoReparacionFiltro)
        {
            var query = _context.Herramienta.AsQueryable();

            if (!string.IsNullOrEmpty(nombreFiltro))
            {
                query = query.Where(h => h.Nombre.Contains(nombreFiltro));
            }
            var herramientasDTO = await query
                .Select(h => new HerramientaDTO()
                {
                    Id = h.Id,
                    Nombre = h.Nombre,
                    Material = h.Material,
                    Fabricante = h.Fabricante,
                    Precio = (double)h.Precio,
                    TiempoReparacion = h.TiempoReparacion
                })
                .ToListAsync();

            if (!string.IsNullOrEmpty(tiempoReparacionFiltro) && int.TryParse(tiempoReparacionFiltro, out int diasFiltro))
            {
                herramientasDTO = herramientasDTO.Where(h =>
                {
                    if (string.IsNullOrEmpty(h.TiempoReparacion)) return false;

                    string soloDigitos = new string(h.TiempoReparacion.Where(char.IsDigit).ToArray());

                    if (int.TryParse(soloDigitos, out int diasReales))
                    {
                        return diasReales <= diasFiltro;
                    }
                    return false;
                }).ToList();
            }

            return Ok(herramientasDTO);
        }


    }
}