# src/AppForSEII2526.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Herramientas/GetSelectFiltradoOferta": {
      "get": {
        "tags": [
          "Herramientas"
        ],
        "operationId": "GetSelectFiltradoOferta",
        "parameters": [
          {
            "name": "fabricante",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "precioMaximo",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaOfertaDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaOfertaDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaOfertaDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Herramientas/GetHerramientaComprar": {
      "get": {
        "tags": [
          "Herramientas"
        ],
        "operationId": "GetHerramientaComprar",
        "parameters": [
          {
            "name": "material",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "precio",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "nombreHerramienta",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaComprarDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaComprarDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaComprarDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Herramientas/GetSelectFiltradoReparacion": {
      "get": {
        "tags": [
          "Herramientas"
        ],
        "operationId": "GetSelectFiltradoReparacion",
        "parameters": [
          {
            "name": "nombreFiltro",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tiempoReparacionFiltro",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Reparacion/GetMostrarReparacionPorId": {
      "get": {
        "tags": [
          "Reparacion"
        ],
        "operationId": "GetMostrarReparacionPorId",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reparacion/CreateReparacion": {
      "post": {
        "tags": [
          "Reparacion"
        ],
        "operationId": "CreateReparacion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReparacionForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReparacionForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReparacionForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReparacionDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HerramientaDTO": {
        "required": [
          "precio"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "double"
          },
          "fabricante": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ReparacionDetailDTO": {
        "type": "object",
        "properties": {
          "nombreCliente": {
            "type": "string",
            "nullable": true
          },
          "apellidoCliente": {
            "type": "string",
            "nullable": true
          },
          "metodoPago": {
            "$ref": "#/components/schemas/TiposMetodoPago"
          },
          "nºTelefono": {
            "type": "string",
            "nullable": true
          },
          "reparacionItem": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReparacionItemDTO"
            },
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fechaEntrega": {
            "type": "string",
            "format": "date-time"
          },
          "fechaRecogida": {
            "type": "string",
            "format": "date-time"
          },
          "precioTotal": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "ReparacionForCreateDTO": {
        "type": "object",
        "properties": {
          "nombreCliente": {
            "type": "string",
            "nullable": true
          },
          "apellidoCliente": {
            "type": "string",
            "nullable": true
          },
          "fechaEntrega": {
            "type": "string",
            "format": "date-time"
          },
          "metodoPago": {
            "$ref": "#/components/schemas/TiposMetodoPago"
          },
          "nºTelefono": {
            "type": "string",
            "nullable": true
          },
          "reparacionItem": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReparacionItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ReparacionItemDTO": {
        "type": "object",
        "properties": {
          "idHerramienta": {
            "type": "integer",
            "format": "int32"
          },
          "nombreHerramienta": {
            "type": "string",
            "nullable": true
          },
          "nombreFabricante": {
            "type": "string",
            "nullable": true
          },
          "tiempoReparacion": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "minimum": 0.01,
            "type": "integer",
            "format": "currency"
          },
          "precioTotal": {
            "type": "number",
            "format": "float"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "TiposMetodoPago": {
        "enum": [
          "TarjetaCredito",
          "PayPal",
          "Efectivo"
        ],
        "type": "string"
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}